pyproject.toml
[project]
name = "mys10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "pydantic>=2.11.3",
    "mcp[cli]>=1.7.0",
    "asyncio>=3.4.3",
    "pymupdf4llm>=0.0.21",
    "pyyaml>=6.0.3",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "bs4>=0.0.2",
    "pillow>=11.2.1",
    "trafilatura[all]>=2.0.0",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "llama-index>=0.12.31",
    "llama-index-embeddings-google-genai>=0.1.0"
]
